fix: add resolvers back

+++ backend/src/main/java/io/papermc/hangar/components/query/QueryHelper.java
@@ -50,7 +50,7 @@ public final class QueryHelper {
         queryBuilder.fields.add(STR."\{parentTable}\{idField} AS \{id}");
 
         queryBuilder.joins.add(STR."JOIN avatars \{parentAlias}avatar ON \{parentAlias}avatar.type = '\{avatarType}' AND \{parentAlias}avatar.subject = \{parentTable}\{idField}::varchar");
-        queryBuilder.resolver.put(parentAlias + "avatarUrl", (r) -> fileService.getAvatarUrl(avatarType, String.valueOf(r.get(id)), (Integer) r.get(avatarVersion)));
+        queryBuilder.resolver.put(parentAlias + "avatarUrl", (r) -> fileService.getAvatarUrl(avatarType, String.valueOf(r.get(id)), r.get(avatarVersion)));
         return EMPTY;
     }
 }
